Change the way testers work to allow for serializable functions in testers.
This allows us to be much more flexible with how we check things. We can use functions
(as long as they are serializable functions, which is a bit limiting but still better than
before) to check things in any way we want. This makes things a bit slower, but allows us
to elegantly express some checks that we need to do.