Make clojail compatible with Clojure version 1.5-1.7
Implement profiles for each version, and add to testall alias to simplify cross
version testing.

Update one test for cross-compatibility. The threading macro `->` was changed
to output a complete result instead of a recursive definition in terms of `->`
in Clojure 1.6. See [this patch](http://dev.clojure.org/jira/browse/CLJ-1121).